+++ test/Microsoft.TestPlatform.Library.IntegrationTests/AppDomainTests.cs
