+++ test/Microsoft.TestPlatform.Library.IntegrationTests/AssemblyMetadataProviderTests.cs
