+++ test/Microsoft.TestPlatform.Library.IntegrationTests/AssemblyPropertiesTests.cs
