+++ test/Microsoft.TestPlatform.Library.IntegrationTests/BannedSymbols.txt
@@ -1,0 +1,3 @@
+﻿M:System.IO.Path.GetTempPath(); Use 'IntegrationTestBase.GetTempPath()' instead
+M:System.Environment.SetEnvironmentVariable(System.String,System.String); Use one of the overload accepting a dictionary of environment variables instead
+M:System.Environment.SetEnvironmentVariable(System.String,System.String,System.EnvironmentVariableTarget); Use one of the overload accepting a dictionary of environment variables instead
