+++ test/Microsoft.TestPlatform.Library.IntegrationTests/DiaSessionTests.cs
