+++ test/Microsoft.TestPlatform.Library.IntegrationTests/MetadataReaderHelperTests.cs
