+++ test/Microsoft.TestPlatform.Library.IntegrationTests/TranslationLayerTests/EventHandler/DiscoveryEventHandler.cs
