+++ test/Microsoft.TestPlatform.TestUtilities/Feature.cs
