+++ test/Microsoft.TestPlatform.TestUtilities/IsExternalInit.cs
