+++ test/Microsoft.TestPlatform.TestUtilities/WrapperCompatibilityDataSource.cs
@@ -6,25 +6,25 @@ using System.Reflection;
 namespace Microsoft.TestPlatform.AcceptanceTests;
 
 /// <summary>
-/// A data source for checking compatibility of changes in testhost with different versions of vstest.console.
-/// This provides the development version of testhost and pairs it with different versions of vstest.console, from Latest to RecentStable.
-/// To give us confidence that we are not breaking customers with recent versions. We don't test with older versions of vstest.console,
-/// because they are less likely to be used by customers that update to latest NET.Test.Sdk.
-/// Because .NET Framework testhost is shipped together with VSTest console, we add only Latest - Latest for .NET Framework testhost.
+/// A data source that checks compatibility of changes in VSTestConsoleWrapper with different versions of vstest.console.
+/// We are testing with all versions of vstest.console, because the wrapper should (ideally) keep backwards compatibility for all shipped
+/// versions of vstest.console.
+/// We also add VSIX to test that the shipment into VS works.
+/// This does NOT test  compatibility of changes in vstest.console, with older versions of the wrapper.
 /// </summary>
-public class TestHostCompatibilityDataSource : TestDataSourceAttribute<RunnerInfo>
+public class WrapperCompatibilityDataSource : CompatibilityDataSourceAttribute
 {
     private readonly CompatibilityRowsBuilder _builder;
 
-    public TestHostCompatibilityDataSource()
+    public WrapperCompatibilityDataSource()
     {
         _builder = new CompatibilityRowsBuilder(
             // runner
-            AcceptanceTestBase.LATEST_TO_RECENT_STABLE,
+            AcceptanceTestBase.LATEST_TO_LEGACY,
             AcceptanceTestBase.DEFAULT_RUNNER_NETFX_AND_NET,
             // host
             AcceptanceTestBase.LATEST,
-            AcceptanceTestBase.DEFAULT_HOST_NETFX_AND_NET,
+            AcceptanceTestBase.DEFAULT_HOST_NET,
             // adapter
             AcceptanceTestBase.LATESTSTABLE,
             AcceptanceTestBase.MSTEST);
@@ -36,21 +36,40 @@ public class TestHostCompatibilityDataSource : TestDataSourceAttribute<RunnerInf
     public bool DebugTestHost { get; set; }
     public bool DebugDataCollector { get; set; }
     public bool DebugStopAtEntrypoint { get; set; }
+    public int JustRow { get; set; } = -1;
 
     public string? BeforeFeature { get; set; }
     public string? AfterFeature { get; set; }
 
+    //public string? BeforeTestHostFeature { get; set; }
+    //public string? AfterTestHostFeature { get; set; }
+
+    public string? BeforeAdapterFeature { get; set; }
+    public string? AfterAdapterFeature { get; set; }
+
     public override void CreateData(MethodInfo methodInfo)
     {
-        _builder.BeforeTestHostFeature = BeforeFeature;
-        _builder.AfterTestHostFeature = AfterFeature;
+        _builder.WithVSIXRunner = true;
+        _builder.WithInProcess = false;
+
+        _builder.BeforeRunnerFeature = BeforeFeature;
+        _builder.AfterRunnerFeature = AfterFeature;
+
+        //_builder.BeforeTestHostFeature = BeforeTestHostFeature;
+        //_builder.AfterTestHostFeature = AfterTestHostFeature;
+
+        _builder.BeforeAdapterFeature = BeforeAdapterFeature;
+        _builder.AfterAdapterFeature = AfterAdapterFeature;
 
         _builder.DebugDataCollector = DebugDataCollector;
         _builder.DebugVSTestConsole = DebugVSTestConsole;
         _builder.DebugTestHost = DebugTestHost;
         _builder.DebugStopAtEntrypoint = DebugStopAtEntrypoint;
 
+        _builder.JustRow = JustRow < 0 ? null : JustRow;
+
         var data = _builder.CreateData();
         data.ForEach(AddData);
     }
 }
+
